requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
pytest==8.3.3
httpx==0.27.2
email-validator==2.2.0
pytest-asyncio==0.20.3
pytest-cov==4.0.0
aiosqlite==0.20.0       